# resources/openapi/partner-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skate & Snow Partner API",
    "version": "1.1.0",
    "description": "API для синхронизации каталога и создания партнёрских заказов. HMAC query канонизируется из сырой строки: каждая key/value пара form-url-decode (включая + как пробел), затем RFC 3986 encode, сортировка по encoded key и value; дубликаты, массивы и пустые значения сохраняются. Каноническая строка: METHOD, /api/partner/v1/path, canonical query без ?, SHA-256 сырого тела, Unix timestamp и уникальный nonce — каждое значение на отдельной строке."
  },
  "servers": [{ "url": "/api/partner/v1", "description": "Текущий сервер" }],
...
  "security": [{ "PartnerKey": [], "PartnerTimestamp": [], "PartnerNonce": [], "PartnerSignature": [] }],
  "paths": {
...
    "/catalog/brands": {
      "get": {
        "tags": ["Catalog"], "summary": "Получить активные бренды", "operationId": "listPartnerBrands", "x-required-scope": "catalog:read",
        "responses": { "200": { "description": "Бренды доступных товаров", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "required": ["success", "data"], "properties": { "success": { "const": true }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Brand" } } } } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/catalog/products": {
      "get": {
        "tags": ["Catalog"], "summary": "Получить товары", "operationId": "listPartnerProducts", "x-required-scope": "catalog:read",
        "parameters": [
          { "name": "category_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "brand_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "q", "in": "query", "schema": { "type": "string", "maxLength": 120 } },
          { "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/PerPage" },
          { "$ref": "#/components/parameters/UpdatedSince" }, { "$ref": "#/components/parameters/Cursor" }
        ],
        "responses": { "200": { "description": "Пагинированный каталог", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductPageResponse" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/catalog/products/{good}": {
      "get": {
        "tags": ["Catalog"], "summary": "Получить товар", "operationId": "getPartnerProduct", "x-required-scope": "catalog:read",
        "parameters": [{ "name": "good", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Товар", "content": { "application/json": { "schema": { "type": "object", "required": ["success", "data"], "properties": { "success": { "const": true }, "data": { "$ref": "#/components/schemas/Product" } } } } } }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "PartnerKey": { "type": "apiKey", "in": "header", "name": "X-Partner-Key", "description": "Key ID, не секрет." },
      "PartnerTimestamp": { "type": "apiKey", "in": "header", "name": "X-Partner-Timestamp", "description": "Текущее Unix-время в секундах. Допуск по умолчанию ±300 секунд." },
      "PartnerNonce": { "type": "apiKey", "in": "header", "name": "X-Partner-Nonce", "description": "Уникальная случайная строка для каждого запроса." },
      "PartnerSignature": { "type": "apiKey", "in": "header", "name": "X-Partner-Signature", "description": "hex(HMAC-SHA256(canonical_string, secret))." }
    },
    "parameters": {
...
      "Page": { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "PerPage": { "name": "per_page", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 24 } },
      "UpdatedSince": { "name": "updated_since", "in": "query", "schema": { "type": "string", "format": "date-time" } },
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string", "maxLength": 512 }, "description": "Непрозрачный cursor из meta.next_cursor; сортировка всегда updated_at,id." }
    },
    "schemas": {
      "Category": { "type": "object", "additionalProperties": false, "required": ["id", "parent_id", "name", "slug", "image_url", "is_active", "sort_order", "updated_at"], "properties": { "id": { "type": "integer" }, "parent_id": { "type": ["integer", "null"] }, "name": { "type": "string" }, "slug": { "type": "string" }, "image_url": { "type": ["string", "null"], "format": "uri" }, "is_active": { "type": "boolean" }, "sort_order": { "type": "integer" }, "updated_at": { "type": ["string", "null"], "format": "date-time" } } },
      "CatalogImage": { "type": "object", "additionalProperties": false, "required": ["id", "url", "alt", "is_main", "sort_order"], "properties": { "id": { "type": "integer" }, "url": { "type": ["string", "null"] }, "alt": { "type": ["string", "null"] }, "is_main": { "type": "boolean" }, "sort_order": { "type": "integer" } } },
      "CatalogAttributeValue": { "type": "object", "additionalProperties": false, "required": ["id", "code", "label", "color"], "properties": { "id": { "type": "integer" }, "code": { "type": "string" }, "label": { "type": "string" }, "color": { "type": ["string", "null"] } } },
      "CatalogAttribute": { "type": "object", "additionalProperties": false, "required": ["id", "code", "name", "value"], "properties": { "id": { "type": "integer" }, "code": { "type": "string" }, "name": { "type": "string" }, "value": { "$ref": "#/components/schemas/CatalogAttributeValue" } } },
      "CatalogProperty": { "type": "object", "additionalProperties": false, "required": ["id", "code", "name", "type", "value", "sort_order"], "properties": { "id": { "type": "integer" }, "code": { "type": "string" }, "name": { "type": "string" }, "type": { "type": "string" }, "value": { "oneOf": [{ "$ref": "#/components/schemas/CatalogAttributeValue" }, { "type": "null" }] }, "sort_order": { "type": "integer" } } },
      "Brand": { "type": "object", "additionalProperties": false, "required": ["id", "name", "slug", "logo_url"], "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "slug": { "type": "string" }, "logo_url": { "type": ["string", "null"], "format": "uri" } } },
      "Dimensions": { "type": "object", "additionalProperties": false, "required": ["length", "width", "height", "weight"], "properties": { "length": { "type": ["number", "null"] }, "width": { "type": ["number", "null"] }, "height": { "type": ["number", "null"] }, "weight": { "type": ["number", "null"] } } },
      "Variation": { "type": "object", "additionalProperties": false, "required": ["id", "sku", "name", "display_name", "attributes", "images", "price", "old_price", "sale_price", "demping_price", "show_demping", "final_price", "discounts", "currency", "available_quantity", "stock_quantity", "remote_stock_quantity", "fast_remote_stock_quantity", "is_active", "is_available", "is_preorder", "purchase_mode", "can_order", "can_preorder", "dimensions", "updated_at"], "properties": { "id": { "type": "integer" }, "sku": { "type": ["string", "null"] }, "name": { "type": ["string", "null"] }, "display_name": { "type": "string" }, "attributes": { "type": "array", "items": { "$ref": "#/components/schemas/CatalogAttribute" } }, "images": { "type": "array", "items": { "$ref": "#/components/schemas/CatalogImage" } }, "price": { "type": "number" }, "old_price": { "type": ["number", "null"] }, "sale_price": { "type": ["number", "null"] }, "demping_price": { "type": ["number", "null"] }, "show_demping": { "type": "boolean" }, "final_price": { "type": "number" }, "discounts": { "type": "object" }, "currency": { "const": "RUB" }, "available_quantity": { "type": "integer", "minimum": 0, "description": "Reservable local stock for this active variation after active reservations; remote stock is excluded." }, "stock_quantity": { "type": "integer", "minimum": 0, "description": "Raw local stock source." }, "remote_stock_quantity": { "type": ["integer", "string", "null"], "description": "Informational remote stock source. Markers such as >10 are preserved and do not enable Partner checkout." }, "fast_remote_stock_quantity": { "type": ["integer", "string", "null"], "description": "Informational fast remote stock source. Markers such as >10 are preserved and do not enable Partner checkout." }, "is_active": { "type": "boolean" }, "is_available": { "type": "boolean", "description": "True only when ordinary Partner checkout can reserve local stock." }, "is_preorder": { "type": "boolean" }, "purchase_mode": { "type": "string", "enum": ["regular", "preorder", "unavailable"] }, "can_order": { "type": "boolean" }, "can_preorder": { "type": "boolean", "description": "Catalog capability flag; preorder checkout is not part of this contract." }, "dimensions": { "$ref": "#/components/schemas/Dimensions" }, "updated_at": { "type": ["string", "null"], "format": "date-time" } } },
      "Product": { "type": "object", "additionalProperties": false, "required": ["id", "name", "slug", "sku", "short_description", "description", "brand", "categories", "images", "properties", "variations", "price", "old_price", "sale_price", "demping_price", "show_demping", "final_price", "discounts", "currency", "available_quantity", "stock_quantity", "remote_stock_quantity", "fast_remote_stock_quantity", "is_active", "is_available", "is_preorder", "purchase_mode", "can_order", "can_preorder", "dimensions", "updated_at"], "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "slug": { "type": "string" }, "sku": { "type": ["string", "null"] }, "short_description": { "type": ["string", "null"] }, "description": { "type": ["string", "null"] }, "brand": { "oneOf": [{ "$ref": "#/components/schemas/Brand" }, { "type": "null" }] }, "categories": { "type": "array", "items": { "$ref": "#/components/schemas/Category" } }, "images": { "type": "array", "items": { "$ref": "#/components/schemas/CatalogImage" } }, "properties": { "type": "array", "items": { "$ref": "#/components/schemas/CatalogProperty" } }, "variations": { "type": "array", "items": { "$ref": "#/components/schemas/Variation" } }, "price": { "type": "number" }, "old_price": { "type": ["number", "null"] }, "sale_price": { "type": ["number", "null"] }, "demping_price": { "type": ["number", "null"] }, "show_demping": { "type": "boolean" }, "final_price": { "type": "number" }, "discounts": { "type": "object" }, "currency": { "const": "RUB" }, "available_quantity": { "type": "integer", "minimum": 0, "description": "Reservable local stock after active reservations. For products with active variations this is the sum across active variations; otherwise it is the product stock." }, "stock_quantity": { "type": "integer", "minimum": 0, "description": "Raw local stock source on the product record." }, "remote_stock_quantity": { "type": ["integer", "string", "null"], "description": "Informational remote stock source. Markers such as >10 are preserved and do not enable Partner checkout." }, "fast_remote_stock_quantity": { "type": ["integer", "string", "null"], "description": "Informational fast remote stock source. Markers such as >10 are preserved and do not enable Partner checkout." }, "is_active": { "type": "boolean" }, "is_available": { "type": "boolean", "description": "True only when at least one reservable local unit is available for ordinary Partner checkout." }, "is_preorder": { "type": "boolean" }, "purchase_mode": { "type": "string", "enum": ["regular", "preorder", "unavailable"] }, "can_order": { "type": "boolean" }, "can_preorder": { "type": "boolean", "description": "Catalog capability flag distinct from ordinary availability; preorder checkout is not part of this contract." }, "dimensions": { "$ref": "#/components/schemas/Dimensions" }, "updated_at": { "type": ["string", "null"], "format": "date-time" } } },
      "CatalogPageMeta": { "type": "object", "additionalProperties": false, "required": ["next_cursor", "sort", "inactive_in_incremental_feed"], "properties": { "next_cursor": { "type": ["string", "null"] }, "sort": { "type": "array", "items": { "type": "string" } }, "inactive_in_incremental_feed": { "type": "boolean" } } },
...
      "ProductPageResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "const": true }, "data": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } } } }, "meta": { "$ref": "#/components/schemas/CatalogPageMeta" } } },
...
      "ErrorDetail": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "retryable": { "type": "boolean" } } },
      "ErrorResponse": { "type": "object", "additionalProperties": false, "required": ["success", "error"], "properties": { "success": { "const": false }, "error": { "$ref": "#/components/schemas/ErrorDetail" } } },
      "Error": { "$ref": "#/components/schemas/ErrorResponse" }
    },
    "responses": {
      "Unauthorized": { "description": "Ошибка HMAC-аутентификации", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "Ключ не имеет требуемого scope", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Ресурс не найден", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
    }
  }
}